+++ src/rex.instrument/src/rex/instrument/interface/draftinstrumentversion.py
@@ -75,7 +75,7 @@ class DraftInstrumentVersion(Extension, Comparable, Displayable, Dictable):
         raise NotImplementedError()
 
     @classmethod
-    def find(cls, offset=0, limit=100, user=None, **search_criteria):
+    def find(cls, offset=0, limit=None, user=None, **search_criteria):
         """
         Returns DraftInstrumentVersions that match the specified criteria.
 
@@ -88,11 +88,13 @@ class DraftInstrumentVersion(Extension, Comparable, Displayable, Dictable):
 
         :param offset:
             the offset in the list of DraftInstrumentVersions to start the
-            return set from (useful for pagination purposes)
+            return set from (useful for pagination purposes); if not specified,
+            defaults to 0
         :type offset: int
         :param limit:
             the maximum number of DraftInstrumentVersions to return (useful for
-            pagination purposes)
+            pagination purposes); if not specified, defaults to ``None``, which
+            means no limit
         :type limit: int
         :param user:
             the User who should have access to the desired