+++ src/rex.instrument/src/rex/instrument/interface/instrumentversion.py
@@ -275,7 +275,7 @@ class InstrumentVersion(Extension, Comparable, Displayable, Dictable):
         raise NotImplementedError()
 
     @classmethod
-    def find(cls, offset=0, limit=100, user=None, **search_criteria):
+    def find(cls, offset=0, limit=None, user=None, **search_criteria):
         """
         Returns InstrumentVersions that match the specified criteria.
 
@@ -288,11 +288,13 @@ class InstrumentVersion(Extension, Comparable, Displayable, Dictable):
 
         :param offset:
             the offset in the list of InstrumentVersions to start the return
-            set from (useful for pagination purposes)
+            set from (useful for pagination purposes); if not specified,
+            defaults to 0
         :type offset: int
         :param limit:
             the maximum number of InstrumentVersions to return (useful for
-            pagination purposes)
+            pagination purposes); if not specified, defaults to ``None``, which
+            means no limit
         :type limit: int
         :param user:
             the User who should have access to the desired InstrumentVersions