+++ src/rex.instrument/src/rex/instrument/interface/user.py
@@ -66,7 +66,7 @@ class User(Extension, Comparable, Displayable, Dictable):
         raise NotImplementedError()
 
     @classmethod
-    def find(cls, offset=0, limit=100, user=None, **search_criteria):
+    def find(cls, offset=0, limit=None, user=None, **search_criteria):
         """
         Returns Users that match the specified criteria.
 
@@ -74,11 +74,12 @@ class User(Extension, Comparable, Displayable, Dictable):
 
         :param offset:
             the offset in the list of Users to start the return set from
-            (useful for pagination purposes)
+            (useful for pagination purposes); if not specified, defaults to 0
         :type offset: int
         :param limit:
             the maximum number of Users to return (useful for pagination
-            purposes)
+            purposes); if not specified, defaults to ``None``, which means
+            no limit
         :type limit: int
         :param user: the User who should have access to the desired Users
         :type user: User